requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==2.0.12
Django==3.2.9
django-admin-rangefilter==0.11.2
django-daterangefilter==1.0.0
idna==3.4
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.18
